+++ libs/test/SmartStoreTest/src/com/salesforce/androidsdk/smartstore/store/SmartStoreLoadExternalStorageTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-present, salesforce.com, inc.
+ * Copyright (c) 2011-present, salesforce.com, inc.
  * All rights reserved.
  * Redistribution and use of this software in source and binary forms, with or
  * without modification, are permitted provided that the following conditions
@@ -24,33 +24,22 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.androidsdk.store;
+package com.salesforce.androidsdk.smartstore.store;
 
 import androidx.test.filters.LargeTest;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-import android.util.Log;
-
 import com.salesforce.androidsdk.analytics.security.Encryptor;
-import com.salesforce.androidsdk.smartstore.store.IndexSpec;
-import com.salesforce.androidsdk.smartstore.store.SmartStore;
 import com.salesforce.androidsdk.smartstore.store.SmartStore.Type;
-import com.salesforce.androidsdk.smartstore.store.SoupSpec;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Test;
+import java.util.Arrays;
+import java.util.Collection;
 import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
 /**
- * More load tests for smartstore - using external storage
+ * Set of tests for the smart store loading numerous and/or large entries and querying them back - using external storage
  */
-@RunWith(AndroidJUnit4.class)
+@RunWith(Parameterized.class)
 @LargeTest
-public class SmartStoreOtherLoadExternalStorageTest extends SmartStoreOtherLoadTest {
-
-    static final int LARGE_BYTES = 512 * 1024;
+public class SmartStoreLoadExternalStorageTest extends SmartStoreLoadTest {
 
     @Override
     protected String getEncryptionKey() {
@@ -62,36 +51,14 @@ public class SmartStoreOtherLoadExternalStorageTest extends SmartStoreOtherLoadT
         store.registerSoupWithSpec(new SoupSpec(soupName, SoupSpec.FEATURE_EXTERNAL_STORAGE), indexSpecs);
     }
 
-    // Test very large payloads for smartstore
-    @Test
-    public void testUpsertLargePayload() throws JSONException {
-        setupSoup(TEST_SOUP, 1, Type.string);
-        JSONObject entry = new JSONObject();
-        for (int i = 0; i < 5; i++) {
-            StringBuilder sb = new StringBuilder();
-            for (int j = 0; j < LARGE_BYTES; j++) {
-                sb.append(i);
-            }
-            entry.put("value_" + i, sb.toString());
-        }
-
-        // Upsert
-        long start = System.currentTimeMillis();
-        store.upsert(TEST_SOUP, entry);
-        long end = System.currentTimeMillis();
-
-        // Log time taken
-        Log.i("SmartStoreLoadTest", "Upserting 5MB+ payload time taken: " + (end - start) + " ms");
-
-        // Verify
-        JSONArray result = store.retrieve(TEST_SOUP, 1L);
-        for (int i = 0; i < 5; i++) {
-            Assert.assertTrue("Value at index " + i + " is incorrect", result.getJSONObject(0).getString("value_" + i).startsWith("" + i));
-        }
-    }
-
-    @Override
-    public void testAlterSoupJSON1Indexing() throws JSONException {
-        // json1 is not compatible with external storage.
+    @Parameterized.Parameters(name = "{0}")
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+//                {"UpsertQuery1StringIndex1000fields1000characters", Type.string, NUMBER_ENTRIES, 1000, 1000, 1}, // to push memory utilization
+                {"UpsertQuery1StringIndex1field20characters", Type.string, NUMBER_ENTRIES, 1, 20, 1},
+                {"UpsertQuery1StringIndex1field1000characters", Type.string, NUMBER_ENTRIES, 1, 1000, 1},
+                {"UpsertQuery1StringIndex10fields20characters", Type.string, NUMBER_ENTRIES, 10, 20, 1},
+                {"UpsertQuery10StringIndexes10fields20characters", Type.string, NUMBER_ENTRIES, 10, 20, 10}
+        });
     }
 }
