+++ libs/test/SmartStoreTest/src/com/salesforce/androidsdk/smartstore/store/SoupSpecTest.java
@@ -24,13 +24,10 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.androidsdk.store;
+package com.salesforce.androidsdk.smartstore.store;
 
-import androidx.test.filters.SmallTest;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import com.salesforce.androidsdk.smartstore.store.SoupSpec;
-
+import androidx.test.filters.SmallTest;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Assert;
